requirements.txt
django
django-distill
markdown
pyyaml
